Added sinon and the ability to mock global props
